pyproject.toml
[tool.poetry]
name = "homework-31"
version = "0.1.0"
description = ""
authors = ["spbproger <zemli2016@yandex.ru>"]
readme = "README.md"
packages = [{include = "homework_31"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
